feat: promote provider tool call stream wrapper
